Seperate actions and reducers
