+++ src/redux/missions/missionActions.js
@@ -2,16 +2,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 
 const GET_MISSIONS = 'GET_MISSIONS';
 
-const missionReducer = (state = [], action) => {
-  switch (action.type) {
-    case 'GET_MISSIONS/fulfilled':
-      return action.payload;
-    default:
-      return state;
-  }
-};
-
-export const fetchMissions = createAsyncThunk(
+const fetchMissions = createAsyncThunk(
   GET_MISSIONS, async () => {
     const response = await fetch('https://api.spacexdata.com/v3/missions');
     const data = await response.json();
@@ -21,11 +12,10 @@ export const fetchMissions = createAsyncThunk(
         mission_id: mission.mission_id,
         mission_name: mission.mission_name,
         description: mission.description,
-        joined: false,
+        reserved: false,
       });
     });
     return dataArr;
   },
 );
-
-export default missionReducer;
+export default fetchMissions;
