+++ src/redux/missions/missionsReducers.js
@@ -1,0 +1,10 @@
+const missionReducer = (state = [], action) => {
+  switch (action.type) {
+    case 'GET_MISSIONS/fulfilled':
+      return action.payload;
+    default:
+      return state;
+  }
+};
+
+export default missionReducer;
